+++ regimes/in/scenarios.go
@@ -4,6 +4,7 @@ package in
 import (
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cbc"
+	"github.com/invopop/gobl/org"
 	"github.com/invopop/gobl/tax"
 )
 
@@ -13,8 +14,8 @@ var invoiceScenarios = &tax.ScenarioSet{
 		// Reverse Charges
 		{
 			Tags: []cbc.Key{tax.TagReverseCharge},
-			Note: &cbc.Note{
-				Key:  cbc.NoteKeyLegal,
+			Note: &tax.ScenarioNote{
+				Key:  org.NoteKeyLegal,
 				Src:  tax.TagReverseCharge,
 				Text: "Reverse Charge",
 			},
@@ -22,8 +23,8 @@ var invoiceScenarios = &tax.ScenarioSet{
 		// Simplified Tax Invoice
 		{
 			Tags: []cbc.Key{tax.TagSimplified},
-			Note: &cbc.Note{
-				Key:  cbc.NoteKeyLegal,
+			Note: &tax.ScenarioNote{
+				Key:  org.NoteKeyLegal,
 				Src:  tax.TagSimplified,
 				Text: "Simplified Tax Invoice",
 			},